Added Check for If in Ghost Dir

+++ updateghost.sh
@@ -7,25 +7,31 @@
 #     echo "This script must be run as root"
 #     exit 1
 # fi
-echo `whoami`
-#Make Tempory Directory and Download Lates Ghost
-mkdir temp
-cd temp
-curl -L -O https://ghost.org/zip/ghost-latest.zip 
-unzip *.zip
-cd ..
+if [ -f config.js ] 
+	then
+	echo `whoami`
+	#Make Tempory Directory and Download Lates Ghost
+	mkdir temp
+	cd temp
+	curl -L -O https://ghost.org/zip/ghost-latest.zip 
+	unzip *.zip
+	cd ..
 
-#Make Backup DB
-cp content/data/ghost.db content/data/ghost_backup.db
+	#Make Backup DB
+	cp content/data/ghost.db content/data/ghost_backup.db
 
-#Copy the new files over
-yes | cp temp/*.md temp/*.js temp/*.json .
-rm -R core
-yes | cp -R temp/core .
-yes | cp -R temp/content/themes/casper content/themes
-npm install --production
+	#Copy the new files over
+	yes | cp temp/*.md temp/*.js temp/*.json .
+	rm -R core
+	yes | cp -R temp/core .
+	yes | cp -R temp/content/themes/casper content/themes
+	npm install --production
 
-#Delete temp folder
-rm -R temp
+	#Delete temp folder
+	rm -R temp
 
-echo "You can now start ghost with npm, forever or whatever else you use"
+	echo "You can now start ghost with npm, forever or whatever else you use"
+else
+	echo "Please cd to your Ghost directory"
+
+fi
